Create PV to MW Calculations
This JavaScript code defines a function called displayDataSummary that calculates and displays various summary statistics about a data set. Here's what it does:

The function takes a single argument, data, which is an array of objects containing data about solar power projects.
The function calculates the total AC and DC capacities of the projects by summing the p_cap_ac and p_cap_dc properties of each object in the data array.
The function calculates the average AC and DC capacities of the projects by taking the mean of the p_cap_ac and p_cap_dc properties of each object in the data array.
The function calculates the estimated total number of solar panels by multiplying the total DC capacity by a constant factor of 5000. This constant factor is assumed to be the number of solar panels needed to generate 1 MW of DC power.
The function creates a visualization variable that selects the #visualization div in the HTML document using D3.js.
The function clears the previous content of the #visualization div by setting its HTML to an empty string.
The function creates a series of p elements using D3.js, and appends them to the visualization div. Each p element displays one of the summary statistics calculated earlier.
Overall, this code provides a simple summary of the solar power project data, displaying the total and average capacities in AC and DC, as well as the estimated total number of solar panels needed to generate the calculated DC capacity. The code can be further customized to display additional summary statistics or to visualize the data in different ways.